streamlined clingo calls
