+++ media/output.json
@@ -196,7 +196,7 @@
           "Class Number": "14553",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
           "Time": "6:00PM - 8:40PM",
-          "Days": "T",
+          "Days": "M",
           "Location": "Remote Learning",
           "Instructor": "Karen Sue Allen"
         }
@@ -311,8 +311,8 @@
         "001": {
           "Class Number": "15200",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "7:30PM - 8:45PM",
-          "Days": "TR",
+          "Time": "4:30PM - 7:10PM",
+          "Days": "W",
           "Location": "Peter Kiewit Institute 259",
           "Instructor": "Ann L Fruhling"
         }
@@ -436,24 +436,24 @@
         "001": {
           "Class Number": "17587",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "6:00PM - 8:40PM",
-          "Days": "M",
+          "Time": "3:00PM - 4:15PM",
+          "Days": "TR",
           "Location": "Peter Kiewit Institute 278",
           "Instructor": "Mahadevan Subramaniam"
         },
         "002": {
           "Class Number": "17588",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "9:00AM - 10:15AM",
-          "Days": "MW",
+          "Time": "1:30PM - 2:20PM",
+          "Days": "T",
           "Location": "Peter Kiewit Institute 278",
           "Instructor": "Mahadevan Subramaniam"
         },
         "003": {
           "Class Number": "17589",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "3:00PM - 4:15PM",
-          "Days": "TR",
+          "Time": "9:30AM - 10:20AM",
+          "Days": "T",
           "Location": "Peter Kiewit Institute 278",
           "Instructor": "Mahadevan Subramaniam"
         }
@@ -497,7 +497,7 @@
         "001": {
           "Class Number": "14260",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "1:30PM - 2:45PM",
+          "Time": "10:30AM - 11:45AM",
           "Days": "TR",
           "Location": "Durham Science Center 254",
           "Instructor": "Dora Velcsov"
@@ -588,7 +588,7 @@
         "001": {
           "Class Number": "15236",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "9:00AM - 10:15AM",
+          "Time": "6:00PM - 7:15PM",
           "Days": "TR",
           "Location": "Peter Kiewit Institute 252",
           "Instructor": "Hesham Hassan Ali"
@@ -619,7 +619,7 @@
           "Class Number": "16070",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
           "Time": "9:00AM - 10:15AM",
-          "Days": "MW",
+          "Days": "TR",
           "Location": "Durham Science Center 104",
           "Instructor": "Fabio Torres Vitor"
         }
@@ -687,7 +687,7 @@
         "002": {
           "Class Number": "16001",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "3:00PM - 4:15PM",
+          "Time": "9:00AM - 10:15AM",
           "Days": "MW",
           "Location": "Peter Kiewit Institute 256",
           "Instructor": "Pei Chi Huang"
@@ -710,8 +710,8 @@
         "820": {
           "Class Number": "14755",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "6:00PM - 8:40PM",
-          "Days": "M",
+          "Time": "4:30PM - 7:10PM",
+          "Days": "W",
           "Location": "Remote Learning",
           "Instructor": "Bill L Sousan"
         }
@@ -725,8 +725,8 @@
         "001": {
           "Class Number": "14389",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "6:00PM - 8:40PM",
-          "Days": "T",
+          "Time": "4:30PM - 7:10PM",
+          "Days": "W",
           "Location": "Peter Kiewit Institute 263",
           "Instructor": "Eric Lundy"
         },
@@ -763,8 +763,8 @@
         "001": {
           "Class Number": "14065",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "3:00PM - 4:15PM",
-          "Days": "TR",
+          "Time": "10:30AM - 11:20AM",
+          "Days": "W",
           "Location": "Peter Kiewit Institute 274",
           "Instructor": "Jorge Fandinno Garcia"
         }
@@ -778,7 +778,7 @@
         "001": {
           "Class Number": "14726",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "10:30AM - 11:45AM",
+          "Time": "3:00PM - 4:15PM",
           "Days": "TR",
           "Location": "Peter Kiewit Institute 276",
           "Instructor": "Xin Zhong"
@@ -823,8 +823,8 @@
         "001": {
           "Class Number": "20110",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "6:00PM - 8:40PM",
-          "Days": "M",
+          "Time": "4:00PM - 5:15PM",
+          "Days": "MW",
           "Location": "Durham Science Center 211A",
           "Instructor": "Mahboub Baccouch"
         }
@@ -853,8 +853,8 @@
         "001": {
           "Class Number": "15948",
           "Date": "Aug 26, 2024 - Dec 20, 2024",
-          "Time": "6:00PM - 8:40PM",
-          "Days": "W",
+          "Time": "1:30PM - 2:45PM",
+          "Days": "TR",
           "Location": "Peter Kiewit Institute 160",
           "Instructor": "Harvey Pe Siy"
         }
